requirements.txt
PyYAML~=6.0
uvicorn~=0.23.2
starlette~=0.27.0
fastapi~=0.104.1
APScheduler~=3.10.4
loguru~=0.7.2
pydantic~=2.4.2